Updated the application to reflect changes made in lecture 27 - specular lighting.
Main features:
- Created a new material class where we can define specular intensity and shininess values.
- Modified the fragment and vertex shaders to get the camera position and place greater specular emphasis on fragments which perfectly reflect to the camera.
- Adjusted model view projection setup for each mesh starting at line 229 in main.cpp.
- Modified pyramid vertices to more accurately reflect specular light.

TO DO:
- Modify cube vertices to better reflect specular lighting.

+++ ModernOpenGLCourseApp/cLight.cpp
@@ -22,8 +22,8 @@ Light::~Light()
 {
 }
 
-void Light::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColourLocation, 
-						GLfloat diffuseIntensityLocation, GLfloat directionLocation)
+void Light::UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
+	GLuint diffuseIntensityLocation, GLuint directionLocation)
 {
 	glUniform3f(ambientColourLocation, mColour.x, mColour.y, mColour.z);
 	glUniform1f(ambientIntensityLocation, mAmbientIntensity);
